+++ apps/modgpt/src/routes/index.ts
@@ -19,20 +19,6 @@ export const protectedRoutes: FastifyPluginCallback = (
   _opts,
   done,
 ) => {
-  if (clerkApplied) {
-    const clerkOptions = {
-      publishableKey: environment.CLERK_PUBLISH_KEY,
-      secretKey: environment.CLERK_SECRET_KEY,
-      jwtKey: environment.CLERK_JWT_KEY,
-    };
-
-    instance.register(clerkPlugin, clerkOptions);
-  } else {
-    if (!clerkApplied)
-      console.warn("No Clerk keys set. Authentication is disabled.");
-    // if (isDevelopment) console.info("ENV set to development");
-  }
-
   getSendChatPath(instance);
   done();
 };
